🔄 refactor(ui): extract dashboard drag-scroll helpers into dedicated module
Move clampDashboardScroll, computeDashboardDragScrollDelta, and
edge-threshold constants from DashboardView.vue into
dashboardDragAutoScroll.ts. Tests import directly from the new
module instead of exercising through DOM pointer events.
